Move core embedding interfaces into `runtime.embed` package.
