+++ runtime/src/main/java/dev/ionfusion/fusion/FusionBool.java
@@ -14,6 +14,7 @@ import com.amazon.ion.IonWriter;
 import com.amazon.ion.ValueFactory;
 import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
 import dev.ionfusion.runtime.base.SourceLocation;
+import dev.ionfusion.runtime.embed.TopLevel;
 import java.io.IOException;
