+++ runtime/src/main/java/dev/ionfusion/fusion/FusionLob.java
@@ -11,6 +11,7 @@ import static dev.ionfusion.fusion.SimpleSyntaxValue.makeSyntax;
 
 import dev.ionfusion.fusion.FusionBool.BaseBool;
 import dev.ionfusion.runtime.base.SourceLocation;
+import dev.ionfusion.runtime.embed.TopLevel;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Arrays;
