+++ runtime/src/main/java/dev/ionfusion/fusion/FusionNull.java
@@ -13,6 +13,7 @@ import com.amazon.ion.ValueFactory;
 import dev.ionfusion.fusion.FusionBool.BaseBool;
 import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
 import dev.ionfusion.runtime.base.SourceLocation;
+import dev.ionfusion.runtime.embed.TopLevel;
 import java.io.IOException;
 
 /**
