+++ runtime/src/main/java/dev/ionfusion/fusion/FusionNumber.java
@@ -24,6 +24,7 @@ import com.amazon.ion.util.IonTextUtils;
 import dev.ionfusion.fusion.FusionBool.BaseBool;
 import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
 import dev.ionfusion.runtime.base.SourceLocation;
+import dev.ionfusion.runtime.embed.TopLevel;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.math.BigInteger;
