+++ runtime/src/main/java/dev/ionfusion/fusion/FusionString.java
@@ -34,6 +34,7 @@ import dev.ionfusion.fusion.FusionBool.BaseBool;
 import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
 import dev.ionfusion.fusion.FusionText.BaseText;
 import dev.ionfusion.runtime.base.SourceLocation;
+import dev.ionfusion.runtime.embed.TopLevel;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.CharBuffer;
