+++ runtime/src/main/java/dev/ionfusion/fusion/FusionSymbol.java
@@ -18,6 +18,7 @@ import com.amazon.ion.util.IonTextUtils;
 import dev.ionfusion.fusion.FusionBool.BaseBool;
 import dev.ionfusion.fusion._private.InternMap;
 import dev.ionfusion.runtime.base.SourceLocation;
+import dev.ionfusion.runtime.embed.TopLevel;
 import java.io.IOException;
 import java.util.Arrays;
