+++ runtime/src/main/java/dev/ionfusion/fusion/FusionTimestamp.java
@@ -35,6 +35,7 @@ import dev.ionfusion.fusion.FusionBool.BaseBool;
 import dev.ionfusion.fusion.FusionNumber.BaseDecimal;
 import dev.ionfusion.fusion.FusionSymbol.BaseSymbol;
 import dev.ionfusion.runtime.base.SourceLocation;
+import dev.ionfusion.runtime.embed.TopLevel;
 import java.io.IOException;
 import java.math.BigDecimal;
