+++ runtime/src/main/java/dev/ionfusion/fusion/FusionVoid.java
@@ -7,6 +7,7 @@ import static dev.ionfusion.fusion.FusionBool.falseBool;
 import static dev.ionfusion.fusion.FusionBool.makeBool;
 
 import dev.ionfusion.fusion.FusionBool.BaseBool;
+import dev.ionfusion.runtime.embed.TopLevel;
 import java.io.IOException;
