+++ runtime/src/main/java/dev/ionfusion/fusion/ModuleBuilderImpl.java
@@ -9,6 +9,7 @@ import com.amazon.ion.IonValue;
 import dev.ionfusion.fusion.Namespace.NsDefinedBinding;
 import dev.ionfusion.fusion._private.doc.model.BindingDoc.Kind;
 import dev.ionfusion.runtime.base.ModuleIdentity;
+import dev.ionfusion.runtime.embed.ModuleBuilder;
 import java.util.Collection;
