+++ runtime/src/main/java/dev/ionfusion/fusion/_Private_CoverageCollectorImpl.java
@@ -4,6 +4,7 @@
 package dev.ionfusion.fusion;
 
 import dev.ionfusion.runtime.base.SourceLocation;
+import dev.ionfusion.runtime.embed.FusionRuntime;
 import java.io.File;
 import java.io.IOException;
 import java.lang.ref.ReferenceQueue;
