+++ runtime/src/main/java/dev/ionfusion/fusion/cli/Document.java
@@ -6,9 +6,9 @@ package dev.ionfusion.fusion.cli;
 import static java.nio.file.Files.isDirectory;
 
 import com.amazon.ion.Timestamp;
-import dev.ionfusion.fusion.FusionRuntime;
 import dev.ionfusion.fusion._private.doc.tool.SiteBuilder;
 import dev.ionfusion.runtime.base.ModuleIdentity;
+import dev.ionfusion.runtime.embed.FusionRuntime;
 import java.io.File;
 import java.io.PrintWriter;
 import java.nio.file.Path;
