+++ runtime/src/main/java/dev/ionfusion/runtime/embed/SandboxBuilder.java
@@ -1,7 +1,10 @@
 // Copyright Ion Fusion contributors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package dev.ionfusion.fusion;
+package dev.ionfusion.runtime.embed;
+
+import dev.ionfusion.fusion.FusionException;
+import dev.ionfusion.fusion.FusionRuntimeBuilder;
 
 /**
  * Constructs {@link TopLevel} namespaces for evaluating code with limited
