+++ runtime/src/test/java/ScriptedTests.java
@@ -5,7 +5,7 @@ import static dev.ionfusion.fusion.TestSetup.makeRuntimeBuilder;
 import static dev.ionfusion.fusion.TestSetup.testRepositoryDirectory;
 import static dev.ionfusion.fusion.TestSetup.testScriptDirectory;
 
-import dev.ionfusion.fusion.FusionRuntime;
+import dev.ionfusion.runtime.embed.FusionRuntime;
 import dev.ionfusion.testing.TreeWalker;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
