+++ runtime/src/test/java/dev/ionfusion/fusion/CoreTestCase.java
@@ -27,6 +27,8 @@ import com.amazon.ion.IonSystem;
 import com.amazon.ion.IonText;
 import com.amazon.ion.IonValue;
 import com.amazon.ion.system.IonSystemBuilder;
+import dev.ionfusion.runtime.embed.FusionRuntime;
+import dev.ionfusion.runtime.embed.TopLevel;
 import dev.ionfusion.testing.StdioTestCase;
 import java.math.BigDecimal;
 import java.math.BigInteger;
