+++ runtime/src/test/java/dev/ionfusion/fusion/FusionValueTest.java
@@ -17,6 +17,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.amazon.ion.IonInt;
 import com.amazon.ion.IonValue;
+import dev.ionfusion.runtime.embed.TopLevel;
 import org.junit.jupiter.api.Test;
 
 public class FusionValueTest
