+++ runtime/src/test/java/dev/ionfusion/fusion/ModuleTest.java
@@ -5,6 +5,8 @@ package dev.ionfusion.fusion;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+
+import dev.ionfusion.runtime.embed.TopLevel;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
