+++ runtime/src/test/java/dev/ionfusion/fusion/RuntimeTest.java
@@ -25,6 +25,7 @@ import com.amazon.ion.system.IonBinaryWriterBuilder;
 import com.amazon.ion.system.SimpleCatalog;
 import dev.ionfusion.runtime.base.ModuleIdentity;
 import dev.ionfusion.runtime.base.SourceName;
+import dev.ionfusion.runtime.embed.ModuleBuilder;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.util.Arrays;
