+++ runtime/src/test/java/dev/ionfusion/fusion/TopLevelTest.java
@@ -5,6 +5,8 @@ package dev.ionfusion.fusion;
 
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import dev.ionfusion.runtime.embed.TopLevel;
 import org.junit.jupiter.api.Test;
 
 /**
